FE-Feat: ScheduleTable 컴포넌트 분리
